Add def.after_on_activate
Currently, as there is no way to extend the `on_activate` function, I propose to add a `after_on_activate()` function that will be run after the usual `mob_activate()`. I understand that fully replacing the `on_activate()` call will damage the API functionality, so in this way, users of the API can extend the activation call with extra initialization routines.